fix: :truck: rename function
